requirements.txt
appnope==0.1.4
asttokens==2.4.1
attrs==23.2.0
comm==0.2.2
coverage==7.5.4
debugpy==1.8.2
decorator==5.1.1
executing==2.0.1
fastjsonschema==2.20.0
iniconfig==2.0.0
ipykernel==6.29.4
ipython==8.26.0
jedi==0.19.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
matplotlib-inline==0.1.7
nbformat==5.10.4
nest-asyncio==1.6.0
numpy==2.0.0
packaging==24.1
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
platformdirs==4.2.2
plotly==5.22.0
pluggy==1.5.0
prompt_toolkit==3.0.47
psutil==6.0.0
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.18.0
pytest==8.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pyzmq==26.0.3
referencing==0.35.1
rpds-py==0.18.1
six==1.16.0
stack-data==0.6.3
tenacity==8.4.2
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
wcwidth==0.2.13
